map in the application-prod properties to override the default.
We want to specify the spring.datasource.password on the prod
host and not use the ones set in the app.jar.

Also set the network type to host so that the mysql running
on the VM (not in docker) is addressible with it only listening
on 127.0.0.1. Otherwise we'd need to do some more advanced
networking configuration.